requirements.txt
# backend/requirements.txt
fastapi
uvicorn[standard]
requests
newspaper3k
beautifulsoup4
lxml               # ← そのまま
lxml_html_clean    # ← ★ 追加
openai
python-dotenv
pydantic-settings
semopy>=2.3,<3.0
pandas
